requirements.txt
streamlit
openai
langchain
python-dotenv
faiss-cpu
tiktoken
